se implementan nuevos atributos en la creacion de la cuenta, se realiza validacion que no permita editar el numero de la cuenta

+++ src/main/java/com/bank/credit_system/entity/AccountDocument.java
@@ -14,7 +14,7 @@ public class AccountDocument {
     @Id
     private String id;
     private String accountType;
-    private Integer accountNumber;
+    private Long accountNumber;
     private Double availableBalance;
     private Double income;
     private Long creationDate;
